+++ src/components/views/ViewDataTable/index.tsx
@@ -3,11 +3,10 @@ import NProgress from 'nprogress';
 import { useRouter } from 'next/router';
 import { DataGridPro, GridColDef, useGridApiRef } from '@mui/x-data-grid-pro';
 import { FormattedMessage, useIntl } from 'react-intl';
-import { FunctionComponent, useContext, useState } from 'react';
+import { FunctionComponent, useState } from 'react';
 import { makeStyles, Snackbar } from '@material-ui/core';
 import { useMutation, useQueryClient } from 'react-query';
 
-import { ConfirmContext } from 'components/ConfirmProvider';
 import createNewView from 'fetching/views/createNewView';
 import deleteViewColumn from 'fetching/views/deleteViewColumn';
 import EmptyView from 'components/views/EmptyView';
@@ -58,7 +57,6 @@ const ViewDataTable: FunctionComponent<ViewDataTableProps> = ({ columns, rows, v
     const { orgId } = router.query;
     const queryClient = useQueryClient();
     const viewId = view.id.toString();
-    const confirm = useContext(ConfirmContext);
 
     const addColumnMutation = useMutation(postViewColumn(orgId as string, viewId), {
         onError: () => {
@@ -176,20 +174,13 @@ const ViewDataTable: FunctionComponent<ViewDataTableProps> = ({ columns, rows, v
     });
 
     const onRowsRemove = () => {
-        confirm.setConfirmProps({
-            actionText: intl.formatMessage({ id: 'misc.views.removeDialog.action' }),
-            onConfirm: () => {
-                setWaiting(true);
-                removeRowsMutation.mutate(selection, {
-                    onSettled: (res) => {
-                        setWaiting(false);
-                        if (res?.failed?.length) setError(VIEW_DATA_TABLE_ERROR.REMOVE_ROWS);
-                    },
-                });
+        setWaiting(true);
+        removeRowsMutation.mutate(selection, {
+            onSettled: (res) => {
+                setWaiting(false);
+                if (res?.failed?.length) setError(VIEW_DATA_TABLE_ERROR.REMOVE_ROWS);
             },
-            title: intl.formatMessage({ id: 'misc.views.removeDialog.title' }),
         });
-        confirm.setOpen(true);
     };
 
     const onViewCreate = () => {
