+++ src/locale/misc/views/en.yml
@@ -64,6 +64,7 @@ removeFromSelection: Remove {numSelected, plural,
     one {1 person}
     other {{numSelected} people}
     } from view
+removeTooltip: Smart search views do not currently support removing rows
 removeDialog:
     action: Are you sure you want to remove these rows from this view?
     title: Remove people from view
